JPA mapping is ready

+++ src/main/java/com/blzr/service/AccountingService.java
@@ -17,7 +17,7 @@ public class AccountingService {
     public Long addActivity(Authority authority, String dateStart, String dateEnd, String volume) throws ParseException, NumberFormatException {
         DateFormat format = new SimpleDateFormat("yyyy-MM-dd");
         connectionService.addActivity(new Activity(authority,
-                format.parse(dateStart), format.parse(dateEnd), Long.valueOf(volume)));
+                format.parse(dateStart), format.parse(dateEnd), Integer.valueOf(volume)));
         return connectionService.countActivity();
     }
 }
